requirements.txt
altair==5.5.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.8
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.6
django-widget-tweaks==1.5.0
dnspython==2.6.1
efficient_apriori==2.0.5
emoji==1.7.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.1.1
fonttools==4.50.0
fqdn==1.5.1
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
kneed==0.8.5
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
monkeylearn==3.6.0
mysql-connector-python==8.0.33
narwhals==1.19.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.4
opencv-python==4.10.0.84
overrides==7.6.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==18.1.0
pycparser==2.21
pydeck==0.9.1
Pygments==2.17.2
pymongo==4.6.3
pyparsing==3.1.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2024.11.6
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.17.1
scikit-learn==1.6.0
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.41.1
tenacity==9.0.0
terminado==0.18.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.1.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
widgetsnbextension==4.0.9
